fix: search for new files failed due to using wrong mount point
When deepin-anything-tool builds the index, it will use the first mount
point of the block device instead of the root mount point. Under certain
conditions, it is impossible to use the file event to update the index.
The solution is to use the root mount point of the block device when
deepin-anything-tool builds the index.

Log: Adapt to V23
Bug: https://pms.uniontech.com/bug-view-144735.html