fix: search fail after partition operation
deepin-anything use the first mount record of the device in
/proc/self/mountinfo to construct the file event path, When some partition
operations occur, the order of multiple mount records of the same device
will change, which leads to an error in the constructed file event path.
The solution is to always pick the mount record for the root of the device.

Log: fix bug