fix: Privileged user mount fuse filesystem stuck
On a system with audit enabled, when a privileged user mounts the fuse
file system, if vfs_monitor obtains information about the fuse file
system (such as through kern_path()), it will cause a system call
deadlock, because the fuse daemon responding to the request Waiting for
mount to complete. The solution is to not get information about mounted
devices by default.

log: fix bug