fix: Fix log files take up a lot of space
Previously, the anything would keep log files generated in the last 30
days. If a large number of logs were generated in a short period of
time, it would take up a lot of space. Let's modify it to only keep the
10 most recent log files.

Log: Fix log files take up a lot of space.