fix: Fix search fail in immutable system
On immutable system, the root partition is mounted at both "/" and
"/sysroot", and subdirectories of the root partition are subsequently
mounted at "/". When an application searches, `anything` uses the
search path to determine the mount point of the root directory of the
corresponding partition. If the search path is "/sysroot/dirname", the
mount point returned in this case should be "/sysroot", not "/" as
returned by `mnt_get_mountpoint()`.

Log: Fix search fail in immutable system
Bug: https://pms.uniontech.com/bug-view-308297.html