feat: Improve performance by saving indexes to tmpfs
The index is saved to tmpfs for quick access by deepin-anything and
other applications. To ensure the validity of the index, a commit
operation is triggered after updating the index and reaching a timeout.
The original index storage directory is solely used for the persistent
storage of the index.

Log: Improve performance