fix: Fail to search after manual partitioning
When a user sets up a separate partition for /home, the root directory
of the partition will be mounted to both /persistent/home and /home. At
this time, the software will incorrectly record the second mount,
causing the event path to be determined incorrectly, and the event will
be filtered and the index cannot be updated.

Log: Fail to search after manual partitioning
Bug: https://pms.uniontech.com/bug-view-314501.html
Bug: https://pms.uniontech.com/bug-view-314963.html
Bug: https://pms.uniontech.com/bug-view-314971.html