fix: Fix anything repeatedly restart when shutdown
The current bug is that the anything-server service exits first,
causing anything-daemon to fail to check the /sys/kernel/vfs_monitor
file, and thus restart before receiving SIGTERM.
The solution is to wait for the stop command from systemd or the file
to reappear without performing additional operations when the
/sys/kernel/vfs_monitor file fails to be checked.

In addition, the program restart logic has been readjusted. If the
error can be recovered by restarting, it will restart, otherwise it
will exit directly.

Log: Fix anything repeatedly restart when shutdown
Bug: https://pms.uniontech.com/bug-view-320627.html